+++ CHANGELOG.md
@@ -3,9 +3,9 @@ TimesheetPlus - Changelog
 
 V1.2
 ----
-* Bugfix: Parsing "c i=<number>" caused segmentation fault 
+* Bugfix: Parsing "c i=<number>" caused segmentation fault
 * Improved removal stability: Attempting to remove more entries than available corrupted timesheet
-
+* Improved stop command: Updating end-time of existing entry now accepts ID argument w/o "i=" prefix
 
 V1.1
 ----